refactor: revert OAuth2 Handler merge
